Feat: Updating while loop within the function def main() at the bottom of bikeshare.py:
Purpose:
The purpose of this code it to ask users if they would like to see stats related to trip duration.

User input:
Yes or no

program output:
if 'yes' the program will output all stats related to trip duration
if 'no' the program will move on to user stats.